requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
groq==0.4.2
httpx==0.27.2
faiss-cpu==1.7.4
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
sentence-transformers==2.3.1
huggingface-hub==0.19.4
transformers==4.36.2
numpy==1.24.3
pandas==1.5.3
PyPDF2==3.0.1
langchain==0.0.354
langchain-community==0.0.13
unstructured==0.10.30
tiktoken==0.5.1
